+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v0.4.0 (in development)
+-----------------------
+- Gave `Attribute` an `as_short_str()` method
+- Displaying an `Attribute` with `{:#}` now produces the short name
+
 v0.3.0 (2026-04-03)
 -------------------
 - Increased MSRV to 1.87